configure HTTP client to trust freeecycle's chain of TLS certs
Freecycle sends a chain of 3 certificates for TLS authentication. These
certificates are valid in the browsers I tried (ff, chrome, safari), and
using curl. But a Go HTTP client, built on my version of OSx, fails to
trust them. It seems that the system certificate pool on OSx (1) does
not contain one of the certs in freecycle's chain, AND (2) has an
interesting certificate trust setting issue that can only be fixed by
toggling the trust settings through the system Keychain (see golang
issue https://github.com/golang/go/issues/24652, and related). This also
seems to be the case in the heroku build environment, where this
application is currently being built/deployed.

I could upload the certificates to my heroku build environment, but that
seems like a one-off solution that would bite me if I change hosts.
Instead, I'm just including the certificate chain with the application
itself, and configuring the http client to trust only that certificate
chain. These are public certificates, so there is no harm in just
passing them along with my application, as far as I could determine
anyway. The downside is that if freecycle changes their certs, the app
will break. At least, if that happens, I will know how to address the
resulting error more quickly.

I am aware that I could also just append these certs to the system pool,
but something just feels cleaner about using only the set of certs that
I know I need for the narrow uses in this app.

This has been an excellent way for me to become more familiar with the
TLS.